hostapp-update-hooks: Use bootcount=0 in the bootcount.env file
Previously, bootcount.env was modified by the u-boot bootcount_ext
driver (never reached production as it was disabled).

We now want to use bootcount.env to store the environment variable
in text format to parse it from u-boot shell commands.

Update the hooks so that instead of removing the file, bootcount=0
is written to it.

Change-type: minor
Changelog-entry: Store bootcount in text format in bootcount.env (affects u-boot only)
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>