balena: Add another healthcheck
Occasionally balena ps and balena info work but the balena daemon is
unable to start a new container. This is usually when something in runc
or containerd is not functioning correctly.

Add a healthcheck to spin up a simple hello-world container as well.

The trade-off here is that there will be a few extra writes to disk
every healthcheck timeout. But there will be a benefit that if runc
or containerd is in a bad state, the healthcheck will fail and systemd
will restart the balena daemon to recover the application container

Fixes #1391

Change-type: patch
Changelog-entry: Add extra healthcheck to balena service. It will spin up a hello-world container as well
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>