hostapp-update: mount mnt-sysroot-inactive when needed
Currently we mount the inactive partition before balena-host and the
rollback services. We don't really need the inactive partition until
it is needed. This slows down boot by a few seconds. Mount the inactive
partition in hostapp-update as we are now preventing the inactive
partition from being mounted by default.

Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>